Stack Marking hack: Start marking from last frame
When a negative push is made by the codegen, it leads to scenarios where
the reachable objects from stack are incorrectly collected by GC. This
hack stashes the old stack top each time there is a negative push and
then when marking marks from the stashed stacktop thus including the
previous frame in the marking process. this may lead to imprecision in
collection but its a tradeoff to avoid lost ovject problem .Exploration
for better and more permanent solution needed.